requirements.txt
google-genai>=0.2.2
pillow>=10.2.0
flask>=3.0.0
werkzeug>=3.0.0
